rename lib from netcompare to jdiff
