+++ .bandit.yml
@@ -3,4 +3,4 @@ skips: []
 # No need to check for security issues in the test scripts!
 exclude_dirs:
   - "./tests/"
-  - "./netcompare/tests/"
+  - "./jdiff/tests/"
