+++ jdiff/operator.py
