+++ jdiff/utils/jmespath_parsers.py
@@ -1,8 +1,8 @@
 """
 jmespath expression parsers and related utilities.
 
-This utility interfaces the custom netcompare jmespath expression with the jmespath library.
-From one expression defined in netcompare, we will derive two expressions: one expression that traverse the json output and get the
+This utility interfaces the custom jdiff jmespath expression with the jmespath library.
+From one expression defined in jdiff, we will derive two expressions: one expression that traverse the json output and get the
 evaluated bit of it, the second will target the reference key relative to the value to evaluate. More on README.md
 """
 import re
